changed directory of app.py and html templates
